Moved meta for workflows down one level
- Removed unused code from microsalt and mip_rna
