+++ cg/meta/workflow/mip_dna.py
@@ -32,7 +32,7 @@ CAPTUREKIT_MAP = {'Agilent Sureselect CRE': 'agilent_sureselect_cre.v1',
 
 
 class AnalysisAPI:
-    """The pipelines are accessed through Trailblazer but cg provides additional conventions and
+    """The workflow is accessed through Trailblazer but cg provides additional conventions and
     hooks into the status database that makes managing analyses simpler"""
 
     def __init__(self, db: Store, hk_api: hk.HousekeeperAPI, scout_api: scoutapi.ScoutAPI,
