+++ cg/meta/workflow/mip_rna.py
@@ -6,36 +6,17 @@ from pathlib import Path
 from typing import List, Any
 from ruamel.yaml import safe_load
 
-from requests.exceptions import HTTPError
-
-from cg.apps import tb, hk, scoutapi, lims
+from cg.apps import tb, hk, lims
 from cg.meta.deliver.mip_dna import MipDnaDeliverAPI
 from cg.apps.pipelines.fastqhandler import BaseFastqHandler
 from cg.store import models, Store
 
 
-COLLABORATORS = ('cust000', 'cust002', 'cust003', 'cust004', 'cust042')
-MASTER_LIST = ('ENDO', 'EP', 'IEM', 'IBMFS', 'mtDNA', 'MIT', 'PEDHEP', 'OMIM-AUTO',
-               'PIDCAD', 'PID', 'SKD', 'NMD', 'CTD', 'IF', 'NEURODEG', 'mcarta')
-COMBOS = {
-    'DSD': ('DSD', 'HYP', 'SEXDIF', 'SEXDET'),
-    'CM': ('CNM', 'CM'),
-    'Horsel': ('Horsel', '141217', '141201'),
-}
-CAPTUREKIT_MAP = {'Agilent Sureselect CRE': 'agilent_sureselect_cre.v1',
-                  'SureSelect CRE': 'agilent_sureselect_cre.v1',
-                  'Agilent Sureselect V5': 'agilent_sureselect.v5',
-                  'SureSelect Focused Exome': 'agilent_sureselect_focusedexome.v1',
-                  'Twist Human core exome v1.3 + Twist Human RefSeq Panel':
-                  'Twist_Target_hg19_RefSeq.bed',
-                  'other': 'agilent_sureselect_cre.v1'}
-
-
 class AnalysisAPI:
-    """The pipelines are accessed through Trailblazer but cg provides additional conventions and
+    """The workflow is accessed through Trailblazer but cg provides additional conventions and
     hooks into the status database that makes managing analyses simpler"""
 
-    def __init__(self, db: Store, hk_api: hk.HousekeeperAPI, scout_api: scoutapi.ScoutAPI,
+    def __init__(self, db: Store, hk_api: hk.HousekeeperAPI,
                  tb_api: tb.TrailblazerAPI, lims_api: lims.LimsAPI, deliver_api: MipDnaDeliverAPI,
                  yaml_loader=safe_load,
                  path_api=Path,
@@ -44,7 +25,6 @@ class AnalysisAPI:
         self.db = db
         self.tb = tb_api
         self.hk = hk_api
-        self.scout = scout_api
         self.lims = lims_api
         self.deliver = deliver_api
         self.yaml_loader = yaml_loader
@@ -124,24 +104,6 @@ class AnalysisAPI:
                 'phenotype': link.status,
                 'expected_coverage': link.sample.application_version.application.sequencing_depth,
             }
-            if sample_data['analysis_type'] in ('tgs', 'wes'):
-                if link.sample.capture_kit:
-                    # set the capture kit from status: key or custom file name
-                    mip_capturekit = CAPTUREKIT_MAP.get(link.sample.capture_kit)
-                    sample_data['capture_kit'] = mip_capturekit or link.sample.capture_kit
-                else:
-                    if link.sample.downsampled_to:
-                        self.LOG.debug(f"{link.sample.name}: downsampled sample, skipping")
-                    else:
-                        try:
-                            capture_kit = self.lims.capture_kit(link.sample.internal_id)
-                            if capture_kit is None or capture_kit == 'NA':
-                                self.LOG.warning(
-                                    f"%s: capture kit not found", link.sample.internal_id)
-                            else:
-                                sample_data['capture_kit'] = CAPTUREKIT_MAP[capture_kit]
-                        except HTTPError:
-                            self.LOG.warning(f"{link.sample.internal_id}: not found (LIMS)")
             if link.mother:
                 sample_data['mother'] = link.mother.internal_id
             if link.father:
@@ -236,33 +198,6 @@ class AnalysisAPI:
 
         fastq_handler.link(case=case, sample=sample, files=files)
 
-    def panel(self, family_obj: models.Family) -> List[str]:
-        """Create the aggregated panel file."""
-        all_panels = self.convert_panels(family_obj.customer.internal_id, family_obj.panels)
-        bed_lines = self.scout.export_panels(all_panels)
-        return bed_lines
-
-    @staticmethod
-    def convert_panels(customer: str, default_panels: List[str]) -> List[str]:
-        """Convert between default panels and all panels included in gene list."""
-        if customer in COLLABORATORS:
-            # check if all default panels are part of master list
-            if all(panel in MASTER_LIST for panel in default_panels):
-                return MASTER_LIST
-
-        # the rest are handled the same way
-        all_panels = set(default_panels)
-
-        # fill in extra panels if selection is part of a combo
-        for panel in default_panels:
-            if panel in COMBOS:
-                for extra_panel in COMBOS[panel]:
-                    all_panels.add(extra_panel)
-
-        # add OMIM to every panel choice
-        all_panels.add('OMIM-AUTO')
-
-        return list(all_panels)
 
     def _get_latest_raw_file(self, family_id: str, tag: str) -> Any:
         """Get a python object file for a tag and a family ."""
@@ -309,12 +244,3 @@ class AnalysisAPI:
                 trending = dict()
 
         return trending
-
-    @staticmethod
-    def is_dna_only_case(case):
-        """returns if all samples of a case has dna application type"""
-
-        for _link in case.links:
-            if _link.sample.application_version.application.analysis_type in 'wts':
-                return False
-        return True
