+++ tests/cli/upload/conftest.py
@@ -9,7 +9,7 @@ from cg.apps.scoutapi import ScoutAPI
 from cg.apps.tb import TrailblazerAPI
 from cg.meta.upload.scoutapi import UploadScoutAPI
 from cg.store import Store
-from cg.meta.mip_dna import AnalysisAPI
+from cg.meta.workflow.mip_dna import AnalysisAPI
 
 
 @pytest.fixture(scope='function', name='base_context')
