+++ tests/cli/workflow/balsamic/conftest.py
@@ -4,7 +4,7 @@ from datetime import datetime
 import pytest
 from cg.apps.balsamic.fastq import BalsamicFastqHandler
 from cg.apps.hk import HousekeeperAPI
-from cg.meta.balsamic import AnalysisAPI
+from cg.meta.workflow.balsamic import AnalysisAPI
 from cg.store import Store, models
